+++ tests/testthat/unit/test-inner_join.R
@@ -15,3 +15,15 @@ test_that("Inner join works.", {
     ans <- paypal %>% start_expr %>% inner_join(website, payment_id = session_id) %>% end_expr
     expect_identical(ans, expected)
 })
+
+test_that("Nesting expressions in inner_join's y works.", {
+    expected <- lhs[lhs[, .(v=max(v)), by=x], on=c("x","v"), nomatch=NULL]
+
+    ans <- lhs %>%
+        start_expr %>%
+        inner_join(nest_expr(.start = FALSE, { .[, .(v = max(v)), by = x] }),
+                   x, v) %>%
+        end_expr
+
+    expect_identical(ans, expected)
+})